+++ source/cbz/muimg.c
@@ -163,7 +163,7 @@ img_load_page(fz_context *ctx, fz_document *doc_, int chapter, int number)
 }
 
 static int
-img_lookup_metadata(fz_context *ctx, fz_document *doc_, const char *key, char *buf, int size)
+img_lookup_metadata(fz_context *ctx, fz_document *doc_, const char *key, char *buf, size_t size)
 {
 	img_document *doc = (img_document*)doc_;
 	if (!strcmp(key, FZ_META_FORMAT))
